+++ test/ppx_irmin/deriver/errors/unsupported_type_object_inherit.ml
@@ -1,0 +1,3 @@
+type y = < y : unit >
+
+type 'a t = < x : unit ; y > [@@deriving irmin]