+++ test/ppx_irmin/deriver/errors/unsupported_type_open_object.ml
@@ -1,0 +1,1 @@
+type 'a t = < .. > [@@deriving irmin]